refactor(components): update dialog-base styles file location and all related files | rk | #4364
