+++ packages/components/src/components/modal/modal-utils.ts
@@ -1,4 +1,4 @@
-import type { Backdrop } from '../../styles/dialog-styles';
+import type { Backdrop } from '../common/dialog-base/dialog-base-styles';
 
 export const MODAL_ARIA_ATTRIBUTES = ['aria-label', 'role'] as const;
 export type ModalAriaAttribute = (typeof MODAL_ARIA_ATTRIBUTES)[number];
